feat: support pointer fields (#64)

+++ example/gen/som/node.group.go
@@ -23,11 +23,14 @@ func (n *group) Query() *query.Group {
 	return query.NewGroup(n.client.db)
 }
 func (n *group) Create(ctx context.Context, group *model.Group) error {
+	if group == nil {
+		return errors.New("the passed node must not be nil")
+	}
 	key := "group"
 	if group.ID != "" {
 		key += ":" + "⟨" + group.ID + "⟩"
 	}
-	data := conv.FromGroup(group)
+	data := conv.FromGroup(*group)
 	raw, err := n.client.db.Create(key, data)
 	if err != nil {
 		return err
@@ -40,7 +43,7 @@ func (n *group) Create(ctx context.Context, group *model.Group) error {
 	if err != nil {
 		return err
 	}
-	*group = *conv.ToGroup(&convNode)
+	*group = conv.ToGroup(convNode)
 	return nil
 }
 func (n *group) Read(ctx context.Context, id string) (*model.Group, bool, error) {
@@ -54,18 +57,22 @@ func (n *group) Read(ctx context.Context, id string) (*model.Group, bool, error)
 	if _, ok := raw.([]any); !ok {
 		raw = []any{raw} // temporary fix
 	}
-	var convNode *conv.Group
+	var convNode conv.Group
 	err = surrealdbgo.Unmarshal(raw, &convNode)
 	if err != nil {
 		return nil, false, err
 	}
-	return conv.ToGroup(convNode), true, nil
+	node := conv.ToGroup(convNode)
+	return &node, true, nil
 }
 func (n *group) Update(ctx context.Context, group *model.Group) error {
+	if group == nil {
+		return errors.New("the passed node must not be nil")
+	}
 	if group.ID == "" {
 		return errors.New("cannot update Group without existing record ID")
 	}
-	data := conv.FromGroup(group)
+	data := conv.FromGroup(*group)
 	raw, err := n.client.db.Update("group:⟨"+group.ID+"⟩", data)
 	if err != nil {
 		return err
@@ -75,10 +82,13 @@ func (n *group) Update(ctx context.Context, group *model.Group) error {
 	if err != nil {
 		return err
 	}
-	*group = *conv.ToGroup(&convNode)
+	*group = conv.ToGroup(convNode)
 	return nil
 }
 func (n *group) Delete(ctx context.Context, group *model.Group) error {
+	if group == nil {
+		return errors.New("the passed node must not be nil")
+	}
 	_, err := n.client.db.Delete("group:⟨" + group.ID + "⟩")
 	if err != nil {
 		return err
